We have taken the Gapminder_All_Time excel file for making the visualization of GapMinder World visualisation on life expectancy rate vs GDP for the various countries in the range of year1900 to 2016.
As per the graph, the countries are divided into 8 regions which are Europe, Asia, Africa, North America, South America, Australia, "Central America and Oceania. So, each region is displayed with some colour coding which are shown in the graph. Countries are labelled with their names and when we bring the mouse of any country it will show the Country name, Population and code of the country. Also, when we click on the bubble of any specific country it shows the region of that countries and all the other countries belongs to that same region will be shown and other countries will get blurred.